+++ PIR/include/Module.hpp
@@ -1050,6 +1050,7 @@ namespace pcit::pir{
 			core::StepAlloc<AtomicRMW, uint32_t> atomic_rmws{};
 
 			core::StepAlloc<MetaLocalVar, uint32_t> meta_local_vars{};
+			core::StepAlloc<MetaParam, uint32_t> meta_params{};
 
 
 			// global values