+++ PIR/src/passes/instCombine.cpp
@@ -867,6 +867,7 @@ namespace pcit::pir::passes{
 			case Expr::Kind::CMPXCHG_SUCCEEDED: return false;
 			case Expr::Kind::ATOMIC_RMW:        return false;
 			case Expr::Kind::META_LOCAL_VAR:    return false;
+			case Expr::Kind::META_PARAM:        return false;
 		}
 
 		evo::debugFatalBreak("Unknown or unsupported Expr::Kind");
@@ -1230,6 +1231,7 @@ namespace pcit::pir::passes{
 			case Expr::Kind::ATOMIC_RMW:        return false;
 
 			case Expr::Kind::META_LOCAL_VAR:    return false;
+			case Expr::Kind::META_PARAM:        return false;
 		}
 
 		evo::debugFatalBreak("Unknown or unsupported Expr::Kind");